use synchronization in SimpleExportActivityProcessor
